[Add] User -> Senpai
+User wird die neue allgemeine Klasse für alle anderen Benutzter auf
Proxer
+Anime und News Benachrichtigungen werden nach getUpdate() in die
Eigenschaft von Senpai
+Eigene Klassen für private Nachrichten, um senden und anschauen zu
vereinfachen
+Eigenschaften für "PMObject" erstellt
