+++ Proxer.API/Community/PrivateMessages/Message.cs
@@ -1,0 +1,12 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Proxer.API.Community.PrivateMessages
+{
+    class Message
+    {
+    }
+}
