+++ Proxer.API/Notifications/NewsNotification.cs
@@ -14,7 +14,7 @@ namespace Proxer.API.Notifications
     /// </summary>
     public class NewsNotification : INotification
     {
-        private readonly User senpai;
+        private readonly Senpai senpai;
         //nachdem getUpdate() einmal aufgerufen wurde werden die Ergebnisse hier drin gespeichert => weniger Auslastung
         private NewsObject[] updateListe;
 
@@ -23,7 +23,7 @@ namespace Proxer.API.Notifications
         /// </summary>
         /// <param name="updateCount"></param>
         /// <param name="senpai"></param>
-        public NewsNotification(int updateCount, User senpai)
+        public NewsNotification(int updateCount, Senpai senpai)
         {
             this.senpai = senpai;
             this.Typ = NotificationType.News;
@@ -44,17 +44,21 @@ namespace Proxer.API.Notifications
         /// <returns></returns>
         public async Task<INotificationObject[]> getUpdates()
         {
-            if (updateListe == null && senpai.LoggedIn)
+            if (this.updateListe == null && senpai.LoggedIn)
             {
                 string lResponse = await HttpUtility.GetWebRequestResponseAsync("http://proxer.me/notifications?format=json&s=news&p=1", senpai.LoginCookies);
                 if (lResponse.StartsWith("{\"error\":0"))
                 {
                     Dictionary<string, List<NewsObject>> lDeserialized = JsonConvert.DeserializeObject<Dictionary<string, List<NewsObject>>>("{" + lResponse.Substring("{\"error\":0,".Length));
-                    
-                    if (this.Count < 15) lDeserialized["notifications"].RemoveRange(this.Count, lDeserialized["notifications"].Count - this.Count);
 
-                    updateListe = lDeserialized["notifications"].ToArray();
-                    return updateListe;
+                    if (this.Count < 15)
+                    {
+                        lDeserialized["notifications"].RemoveRange(this.Count, lDeserialized["notifications"].Count - this.Count);
+                        senpai.News.RemoveRange(this.Count, lDeserialized["notifications"].Count - this.Count);
+                    }
+
+                    this.updateListe = lDeserialized["notifications"].ToArray();
+                    return this.updateListe;
                 }
                 else
                 {
@@ -64,7 +68,7 @@ namespace Proxer.API.Notifications
             else
             {
                 //kann auch null sein
-                return updateListe;
+                return this.updateListe;
             }
         }
     }
