+++ Proxer.API/Proxer.API.csproj
@@ -59,6 +59,8 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Community\PrivateMessages\Conference.cs" />
+    <Compile Include="Community\PrivateMessages\Message.cs" />
     <Compile Include="EventArgs\NotificationEventArgs.cs" />
     <Compile Include="Notifications\NotificationObjects\NotificationObjectType.cs" />
     <Compile Include="Notifications\NotificationType.cs" />
@@ -71,8 +73,9 @@
     <Compile Include="Notifications\NotificationObjects\PMObject.cs" />
     <Compile Include="Notifications\NotificationObjects\AnimeMangaUpdateObject.cs" />
     <Compile Include="Notifications\PMNotification.cs" />
-    <Compile Include="Notifications\UpdateNotification.cs" />
+    <Compile Include="Notifications\AnimeMangaNotification.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Senpai.cs" />
     <Compile Include="User.cs" />
     <Compile Include="Utility\HttpUtility.cs" />
     <Compile Include="Utility\Utility.cs" />
